I cleaned up the app via Noel's changes (Below). It is now ready for publishing. I just need the go ahead from Noel.
-The App is called VTG (on the screen it reads VTC to me)
-If you can is it possible on the first page to change "View PDF"
to just download.
-On the inflator is it possible to put a donate button that goes
 to my paypal?  the address is kissena@hotmail.com -  this can
replace help for now.
-Change the pictures on the main screens. (the ones I sent you)
-For the youtube link, the ones that show all four shapes you can
link to the VTG #1/3 video, the ones with the two colors can go
to VTG#2/3 and the ones with both shapes and colors the VTG#3
video.  We tried to shoot all 40 but it was a little dicey,
that will be in the update.

Signed-off-by: Kevin  Nennig <trinerd213@gmail.com>