This update is in the backend concerning the db file. It has been updated to become more modulated for expansion of the application. Nothing on the front end has been changed. The code is now using db_v2.csv and the file SingletonPoiMoveMap was majorly overhauled. The parser will now be able to handle other poi sets and not just 3:1::1:3.
When updating the db_v2.csv, the naming scheme must be the exact same as
other fields already added. And when adding fields to the PoiMove class,
the naming scheme needs to be followed.

Further updates will occur on the front end, iniclude the "revealing" of
lite vs pro versions of the app.

Signed-off-by: knennigtri <knennig213@gmail.com>